feat add webserver with routing & graceful shutdown
This needs to be factored out better (and in general), but it's a rough
and ready HTTP server with graceful shutdown, as well as routing to
different handlers depending on paths.

It's basically a demonstration toy, and it could be refactored to be
quite nice... After some tests are written!